% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make.qq.manhattan.R
\name{make.qq.manhattan}
\alias{make.qq.manhattan}
\title{Read in association results and create corresponding
QQ and possibly Manhattan plots}
\usage{
make.qq.manhattan(
  filename.pvals,
  trait.name,
  filename.prev.hits = NA,
  filename.novel.hits = NA,
  output.filestem = "output",
  gc.correct = FALSE,
  write.locus.labels = TRUE,
  rsid.header = "SNP",
  pval.header = "PVAL",
  chr.header = "CHR",
  pos.header = "POS",
  background.snp.color.one = "darkgrey",
  background.snp.color.two = "lightgrey",
  known.locus.color = "blue",
  novel.locus.color = "red",
  gws.threshold = NA,
  truncate.p = NA
)
}
\arguments{
\item{filename.pvals}{character vector; name of
input file with minimal association data. the file
should have a header line. the file can be
compressed with gzip (.gz) or bzip2 (.bz2)}

\item{trait.name}{character vector; name of
phenotype being analyzed for pretty printing}

\item{filename.prev.hits}{character vector; name
of file containing known associated variants. this is
conceptually expected to be variants from prior
publications, as opposed to the current study
for which the plot is being created. can be NA
if no such file is present}

\item{filename.novel.hits}{character vector; name
of file containing novel associated variants. this
is conceptually expected to be variants from the
current study for which the plot is being created.
can be NA if no such file is present}

\item{output.filestem}{character vector; name
prefix for all output files (QQ and/or Manhattan
plot)}

\item{gc.correct}{logical; whether to apply
GC correction to p-values. GC (genomic control)
correction adjusts the 1-df Chi square distribution
of input p-values such that the median of the
distribution is precisely the median of the
uniform random p-value distribution on the units
interval}

\item{write.locus.labels}{logical; whether to
render labels for known/novel signal highlights
in the Manhattan plot}

\item{rsid.header}{character vector; name of
column in association results corresponding to
variant IDs}

\item{pval.header}{character vector; name of
column in association results corresponding to
p-values}

\item{chr.header}{character vector; name of
column in association results corresponding to
chromosome}

\item{pos.header}{character vector; name of
column in association results corresponding to
physical position}

\item{background.snp.color.one}{character
vector; darker color for emphasizing alternating
chromosomes in Manhattan plot}

\item{background.snp.color.two}{character
vector; lighter color for emphasizing alternating
chromosomes in Manhattan plot}

\item{known.locus.color}{character vector; name
of color for highlighting previously known loci}

\item{novel.locus.color}{character vector; name
of color for highlighting novel associated loci}

\item{gws.threshold}{numeric; p-value threshold
to use for genome-wide significance threshold;
defaults to 5e-8. if no threshold should be rendered,
set to NA}

\item{truncate.p}{numeric; p-value threshold
below which to arbitrarily override observed
p-values. can be used to make more manageable plots
when extremely small p-values are present. if no
truncation is required, either set to NA or a
small enough value such that no variants are
affected}
}
\description{
This is the primary entry point for the legacy code.
The user provides an analysis file with columns at
least for: chromosome, physical position, SNP ID
(often rsID), and p-value. These are processed and
possibly cross-referenced with externally prepared
lists of known signals. Then, depending on user
specification, a QQ plot and Manhattan plot (with
known signals highlighted) are emitted.
}
