% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/themes.R
\docType{data}
\name{my.theme}
\alias{my.theme}
\title{Apply basic ggplot2 theme settings}
\format{
An object of class \code{theme} (inherits from \code{gg}) of length 93.
}
\usage{
my.theme
}
\description{
A simple utility definition of some straightforward
plot settings. Nothing too interesting
}
\keyword{datasets}
