% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render.manhattan.plot.R
\name{render.manhattan.plot}
\alias{render.manhattan.plot}
\title{Given input data and a variety of settings,
create a Manhattan plot and save it to file.}
\usage{
render.manhattan.plot(
  raw.data,
  pos.col.header,
  pval.col.header,
  all.colors,
  prev.loci.color,
  novel.loci.color,
  snp.color.one,
  snp.color.two,
  gws.threshold,
  chr.center,
  chr.labels,
  y.max,
  filename.prev.hits,
  filename.novel.hits,
  write.locus.labels,
  snp.labels,
  snp.label.x.positions,
  snp.label.y.positions,
  output.filestem
)
}
\arguments{
\item{raw.data}{data.frame; input data for
Manhattan plot with appropriate transformations}

\item{pos.col.header}{character vector; header
name for column of transformed physical positions
in input data}

\item{pval.col.header}{character vector; header
name for column of transformed p-values in input
data}

\item{all.colors}{character vector; color settings
for each row of input data}

\item{prev.loci.color}{character vector; color
setting for known loci highlighting}

\item{novel.loci.color}{character vector; color
setting for novel loci highlighting}

\item{snp.color.one}{character vector;
darker color for emphasizing alternating
chromosomes in Manhattan plot}

\item{snp.color.two}{character vector;
lighter color for emphasizing alternating
chromosomes in Manhattan plot
gws.threshold numeric; p-value threshold
to use for genome-wide significance threshold;
defaults to 5e-8. if no threshold should be rendered,
set to NA
chr.center numeric vector; location of chromosome
centers along the x-axis for labeling purposes
chr.labels character vector; labels to render along
x-axis (e.g.: "1", "2", "23", "XY" if you want, etc.)
y.max numeric; override value to set as maximum yintercept
position in output plot. should exceed valid space
of observed points by some amount for aesthetically
pleasing plotting
filename.prev.hits character vector; name of file
of known signals, or NA. this is only used to control
whether other data are processed as if these signals
exist; the file itself is not read. this is a candidate
for removal from the function interface
filename.novel.hits character vector; name of file
of novel signals, or NA. this is only used to control
whether other data are processed as if these signals
exist; the file itself is not read. this is a candidate
for removal from the function interface
write.locus.labels logical; whether to
render labels for known/novel signal highlights
snp.labels character vector; labels to be applied
to highlighted loci in plot. probably should be
renamed to "loci.labels" for consistency
snp.label.x.positions numeric vector; x coordinates
for loci labels. should probably be renamed to
"loci.label.x.positions" for consistency
snp.label.y.positions numeric vector; y coordinates
for loci labels. should probably be renamed to
"loci.label.y.positions" for consistency
output.filestem character vector; prefix for output
plot, to which will be appended ".manhattan.jpg"}
}
\description{
This function only handles direct interactions
with \code{ggplot2}; it doesn't do much in the way of
input data processing. most of that is handled
by other library functions. the description
of this process will be expanded in later versions.
}
